阿哈哈哈 基本可用

+++ src/main/java/com/dawei/boot/boothelloword/graphql/entry/AuthorInfo.java
@@ -8,6 +8,6 @@ import lombok.Data;
 @Data
 public class AuthorInfo {
 
-	private long authorId;
+	private int authorId;
 	private String authorName;
 }
